backend/python_service/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
python-dotenv==1.0.0
openai==1.3.0
pydantic==2.4.2
httpx==0.24.1
motor==3.3.2
pymongo==4.6.3